Article crud done

+++ resources/js/web/Admin/pages/Categories.vue
@@ -163,8 +163,20 @@ export default {
         .setAttribute("class", "modal fade hide");
     },
     destroy(id) {
-      
-      this.$store.dispatch("category/destroy", id);
+      Swal.fire({
+        title: "Are you sure?",
+        text: "You won't be able to revert this!",
+        type: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Yes, delete it!"
+      }).then(result => {
+        if (result.value) {
+          this.$store.dispatch("category/destroy", id);
+          Swal.fire("Deleted!", "Your file has been deleted.", "success");
+        }
+      });
     }
   },
   mounted() {
